requirements.txt
sdfcad
jupyterlab
pyvista

manifold3d
trimesh
pyglet<2
pyrender